feat(parent-auth): liff-login 成功時多發 30 天 refresh token cookie
DB 寫入 ParentRefreshToken row（family 起點 parent_token_id=NULL）。
記錄 user_agent / ip 供觀測，不參與決策。

修正：id 欄位改 Integer（SQLite autoincrement 相容性）。

+++ alembic/versions/20260503_b7c8d9e0f1g2_create_parent_refresh_tokens.py
@@ -34,7 +34,7 @@ def upgrade() -> None:
 
     op.create_table(
         "parent_refresh_tokens",
-        sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
         sa.Column(
             "user_id",
             sa.BigInteger,
@@ -51,7 +51,7 @@ def upgrade() -> None:
         ),
         sa.Column(
             "parent_token_id",
-            sa.BigInteger,
+            sa.Integer,
             sa.ForeignKey("parent_refresh_tokens.id", ondelete="SET NULL"),
             nullable=True,
         ),
